Profile: improve friendViewController and UserCell

+++ joyyios/JYProfileCardView.h
@@ -11,8 +11,8 @@
 @interface JYProfileCardView : UIView
 
 @property (nonatomic) JYUser *user;
-@property (nonatomic) uint32_t friendCount;
-@property (nonatomic) uint32_t inviteCount;
-@property (nonatomic) uint32_t winkCount;
+@property (nonatomic) uint64_t friendCount;
+@property (nonatomic) uint64_t inviteCount;
+@property (nonatomic) uint64_t winkCount;
 
 @end
